# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "answer_mode": {
            "anyOf": [
              {
                "enum": [
                  "standard",
                  "theological",
                  "pastoral",
                  "beginner",
                  "kids"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Mode"
          },
          "chatbot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chatbot Id"
          },
          "participant_category": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Category"
          },
          "participant_name": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Name"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "query"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "closing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing"
          },
          "featured_malssum": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeaturedMalssum"
              },
              {
                "type": "null"
              }
            ]
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "persona_overridden": {
            "default": false,
            "title": "Persona Overridden",
            "type": "boolean"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          },
          "suggested_followups": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Followups"
          }
        },
        "required": [
          "answer",
          "sources",
          "session_id",
          "message_id"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "ChatbotConfigCreate": {
        "properties": {
          "chatbot_id": {
            "title": "Chatbot Id",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "persona_name": {
            "default": "",
            "title": "Persona Name",
            "type": "string"
          },
          "search_tiers": {
            "$ref": "#/components/schemas/SearchTiersConfig"
          },
          "streaming_enabled": {
            "default": true,
            "title": "Streaming Enabled",
            "type": "boolean"
          },
          "system_prompt": {
            "default": "",
            "title": "System Prompt",
            "type": "string"
          }
        },
        "required": [
          "chatbot_id",
          "display_name"
        ],
        "title": "ChatbotConfigCreate",
        "type": "object"
      },
      "ChatbotConfigResponse": {
        "properties": {
          "chatbot_id": {
            "title": "Chatbot Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "persona_name": {
            "title": "Persona Name",
            "type": "string"
          },
          "search_tiers": {
            "$ref": "#/components/schemas/SearchTiersConfig"
          },
          "streaming_enabled": {
            "title": "Streaming Enabled",
            "type": "boolean"
          },
          "suggested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested At"
          },
          "suggested_questions": {
            "items": {
              "type": "string"
            },
            "title": "Suggested Questions",
            "type": "array"
          },
          "system_prompt": {
            "title": "System Prompt",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "chatbot_id",
          "display_name",
          "description",
          "system_prompt",
          "persona_name",
          "search_tiers",
          "is_active",
          "streaming_enabled",
          "created_at",
          "updated_at"
        ],
        "title": "ChatbotConfigResponse",
        "type": "object"
      },
      "ChatbotConfigUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name"
          },
          "search_tiers": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchTiersConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "streaming_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Streaming Enabled"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          }
        },
        "title": "ChatbotConfigUpdate",
        "type": "object"
      },
...
      "FeaturedMalssum": {
        "description": "레드팀 시연 — 답변 화면에 무작위로 곁들이는 큐레이션 말씀 1개.\n\n의미 검색이 아니라 사람이 추린 목록에서 random 선택. 매 응답마다 새로\n계산되며 캐시 payload 에는 포함하지 않는다 (항상 fresh).",
        "properties": {
          "category": {
            "default": "",
            "title": "Category",
            "type": "string"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "volume": {
            "default": "",
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "FeaturedMalssum",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedResponse_ChatbotConfigResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatbotConfigResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[ChatbotConfigResponse]",
        "type": "object"
      },
...
      "ReactionAggregate": {
        "description": "단일 message_id 의 반응 카운트 집계 — UI badge 용.",
        "properties": {
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "save": {
            "default": 0,
            "title": "Save",
            "type": "integer"
          },
          "thumbs_down": {
            "default": 0,
            "title": "Thumbs Down",
            "type": "integer"
          },
          "thumbs_up": {
            "default": 0,
            "title": "Thumbs Up",
            "type": "integer"
          }
        },
        "required": [
          "message_id"
        ],
        "title": "ReactionAggregate",
        "type": "object"
      },
...
      "ReactionRequest": {
        "description": "POST /api/chat/messages/{id}/reaction body.\n\nB2 보안: ``user_session_id`` 는 더 이상 클라이언트 입력이 아니다 —\n서버가 HttpOnly cookie ``tw_anon_session`` 으로 발급/관리한다.",
        "properties": {
          "kind": {
            "enum": [
              "thumbs_up",
              "thumbs_down",
              "save"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ReactionRequest",
        "type": "object"
      },
      "ReactionResponse": {
        "description": "단일 반응 응답.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "thumbs_up",
              "thumbs_down",
              "save"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "user_session_id": {
            "title": "User Session Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "message_id",
          "user_session_id",
          "kind",
          "created_at"
        ],
        "title": "ReactionResponse",
        "type": "object"
      },
      "ReactionToggleResponse": {
        "description": "토글 결과 — added 면 reaction 포함, removed 면 없음.",
        "properties": {
          "action": {
            "enum": [
              "added",
              "removed"
            ],
            "title": "Action",
            "type": "string"
          },
          "reaction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReactionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action"
        ],
        "title": "ReactionToggleResponse",
        "type": "object"
      },
...
      "SearchTierSchema": {
        "description": "개별 검색 티어 설정.",
        "properties": {
          "min_results": {
            "default": 3,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Min Results",
            "type": "integer"
          },
          "score_threshold": {
            "default": 0.1,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score Threshold",
            "type": "number"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "sources"
        ],
        "title": "SearchTierSchema",
        "type": "object"
      },
      "SearchTiersConfig": {
        "description": "search_tiers JSONB 구조.",
        "properties": {
          "dictionary_enabled": {
            "default": false,
            "title": "Dictionary Enabled",
            "type": "boolean"
          },
          "multiturn_enabled": {
            "default": true,
            "title": "Multiturn Enabled",
            "type": "boolean"
          },
          "query_rewrite_enabled": {
            "default": false,
            "title": "Query Rewrite Enabled",
            "type": "boolean"
          },
          "raw_rag_only": {
            "default": false,
            "title": "Raw Rag Only",
            "type": "boolean"
          },
          "rerank_enabled": {
            "default": false,
            "title": "Rerank Enabled",
            "type": "boolean"
          },
          "search_mode": {
            "default": "cascading",
            "enum": [
              "cascading",
              "weighted"
            ],
            "title": "Search Mode",
            "type": "string"
          },
          "tiers": {
            "items": {
              "$ref": "#/components/schemas/SearchTierSchema"
            },
            "title": "Tiers",
            "type": "array"
          },
          "weighted_sources": {
            "items": {
              "$ref": "#/components/schemas/WeightedSourceSchema"
            },
            "title": "Weighted Sources",
            "type": "array"
          }
        },
        "title": "SearchTiersConfig",
        "type": "object"
      },
...
      "Source": {
        "properties": {
          "chunk_id": {
            "default": "",
            "title": "Chunk Id",
            "type": "string"
          },
          "cited_phrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cited Phrase"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "volume",
          "text",
          "score"
        ],
        "title": "Source",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TrueWords RAG Platform",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/chatbot-configs": {
      "get": {
        "description": "챗봇 설정 목록 조회 (페이지네이션, created_at DESC).",
        "operationId": "list_all_configs_admin_chatbot_configs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ChatbotConfigResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List All Configs",
        "tags": [
          "admin-chatbot"
        ]
      },
      "post": {
        "operationId": "create_config_admin_chatbot_configs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Config",
        "tags": [
          "admin-chatbot"
        ]
      }
    },
    "/admin/chatbot-configs/{config_id}": {
      "get": {
        "description": "챗봇 설정 단건 조회.",
        "operationId": "get_config_admin_chatbot_configs__config_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Config Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config",
        "tags": [
          "admin-chatbot"
        ]
      },
      "put": {
        "operationId": "update_config_admin_chatbot_configs__config_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Config Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config",
        "tags": [
          "admin-chatbot"
        ]
      }
    },
...
    "/api/chat/messages/{message_id}/reaction": {
      "post": {
        "description": "단일 반응 토글 (P1-A).\n\n인증/세션:\n    - HttpOnly cookie ``tw_anon_session`` 에서 user_session_id 를 읽어 사용한다.\n    - 쿠키 부재 시 서버가 즉시 발급해 Set-Cookie 헤더로 전달.\n    - 클라이언트는 user_session_id 를 직접 제공할 수 없다.\n\nRate limit:\n    - IP 기준. 초과 시 ``RateLimitExceededError``.\n\n토글 시맨틱 (atomic):\n    - 동일 (message_id, anon_session, kind) 가 있으면 제거.\n    - 없으면 INSERT 시도 → ``IntegrityError`` 발생 시 (race) 이미 존재로\n      판단해 \"removed\" 반환 (service 내부에서 처리).",
        "operationId": "toggle_reaction_api_chat_messages__message_id__reaction_post",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionToggleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Reaction",
        "tags": [
          "chat-reactions"
        ]
      }
    },
    "/api/chat/messages/{message_id}/reactions": {
      "get": {
        "description": "단일 message_id 의 반응 카운트 (UI badge 용). 인증 불필요 (집계만).",
        "operationId": "get_aggregate_api_chat_messages__message_id__reactions_get",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionAggregate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Aggregate",
        "tags": [
          "chat-reactions"
        ]
      }
    },
...
    "/chat": {
      "post": {
        "description": "RAG 기반 채팅 응답. 에러는 글로벌 exception_handler가 처리.\n\n로그인 상태면 세션을 사용자에게 귀속(user_id) → 대화 기록 목록 조회 대상.",
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}